Final decimal fix

+++ SharedContracts/SharedContracts/DTOs/Identity/Requests/CreateAddressRequest.cs
@@ -38,10 +38,10 @@ namespace PureDelivery.Shared.Contracts.DTOs.Identity.Requests
         public string DeliveryInstructions { get; set; }
 
         [Range(-90, 90, ErrorMessage = "Latitude must be between -90 and 90")]
-        public decimal? Latitude { get; set; }
+        public double? Latitude { get; set; }
 
         [Range(-180, 180, ErrorMessage = "Longitude must be between -180 and 180")]
-        public decimal? Longitude { get; set; }
+        public double? Longitude { get; set; }
 
         [Required(ErrorMessage = "IsDefault is required")]
         public bool IsDefault { get; set; }
